Fix failing test, rename types file
