+++ src/enqueued-task.ts
@@ -1,4 +1,4 @@
-import type { EnqueuedTaskObject } from "./types/types.js";
+import type { EnqueuedTaskObject } from "./types.js";
 
 class EnqueuedTask {
   taskUid: EnqueuedTaskObject["taskUid"];
