+++ src/index.ts
@@ -1,4 +1,4 @@
-export * from "./types/types.js";
+export * from "./types.js";
 export * from "./errors/index.js";
 export * from "./indexes.js";
 export * from "./enqueued-task.js";
