+++ src/meilisearch.ts
@@ -32,8 +32,8 @@ import type {
   BatchesResults,
   BatchesQuery,
   MultiSearchResponseOrSearchResponse,
-} from "./types/types.js";
-import { ErrorStatusCode } from "./types/types.js";
+} from "./types.js";
+import { ErrorStatusCode } from "./types.js";
 import { HttpRequests } from "./http-requests.js";
 import { TaskClient, type Task } from "./task.js";
 import { EnqueuedTask } from "./enqueued-task.js";
