+++ tests/keys.test.ts
@@ -1,6 +1,6 @@
 import { expect, test, describe, beforeEach, afterAll } from "vitest";
 import { MeiliSearch } from "../src/index.js";
-import { ErrorStatusCode } from "../src/types/types.js";
+import { ErrorStatusCode } from "../src/types.js";
 import {
   clearAllIndexes,
   config,
