+++ tests/pagination.test.ts
@@ -6,7 +6,7 @@
   afterAll,
   beforeAll,
 } from "vitest";
-import { ErrorStatusCode } from "../src/types/types.js";
+import { ErrorStatusCode } from "../src/types.js";
 import {
   clearAllIndexes,
   config,
