+++ tests/typed_search.test.ts
@@ -6,7 +6,7 @@ import {
   expect,
   test,
 } from "vitest";
-import { ErrorStatusCode, type SearchResponse } from "../src/types/types.js";
+import { ErrorStatusCode, type SearchResponse } from "../src/types.js";
 import { EnqueuedTask } from "../src/enqueued-task.js";
 import {
   clearAllIndexes,
